requirements.txt
streamlit
langchain
google-generativeai
streamlit-chat

